+++ lua/obsidian/completion/sources/blink/util.lua
@@ -1,5 +1,17 @@
 local M = {}
 
+---Safe version of vim.str_utfindex
+---@param text string
+---@param vimindex integer|nil
+---@return integer
+local to_utfindex = function(text, vimindex)
+  vimindex = vimindex or #text + 1
+  if vim.fn.has "nvim-0.11" == 1 then
+    return vim.str_utfindex(text, "utf-16", math.max(0, math.min(vimindex - 1, #text)))
+  end
+  return vim.str_utfindex(text, math.max(0, math.min(vimindex - 1, #text)))
+end
+
 ---Generates the completion request from a blink context
 ---@param context blink.cmp.Context
 ---@return obsidian.completion.sources.base.Request
@@ -8,6 +20,8 @@ M.generate_completion_request_from_editor_state = function(context)
   local col = context.cursor[2] + 1
   local cursor_before_line = context.line:sub(1, col - 1)
   local cursor_after_line = context.line:sub(col)
+
+  local character = to_utfindex(context.line, col)
 
   return {
     context = {
@@ -18,6 +32,7 @@ M.generate_completion_request_from_editor_state = function(context)
         row = row,
         col = col,
         line = row + 1,
+        character = character,
       },
     },
   }